tests pass after native refactor
