+++ lib/cluster/cluster_master.js
@@ -9,7 +9,6 @@ module.exports = function (context) {
     const logger = context.foundation.makeLogger({ module: 'cluster_master' });
     const clusterConfig = context.sysconfig.teraslice;
     const parseError = require('../utils/error_utils').parseError;
-    const events = context.foundation.getEventEmitter();
 
     // Initialize the HTTP service for handling incoming requests.
     const app = require('express')();
@@ -22,32 +21,11 @@ module.exports = function (context) {
 
     const assetServiceOnline = assetServicelistener();
 
-    require('./services/execution.js')(context)
+    // ignore process calls, this is handled by ipc messages from node master
+    messaging.register('process:SIGTERM', () => {});
+    messaging.register('process:SIGINT', () => {});
 
-        .then((clusterService) => {
-            messaging.initialize({ server });
-            logger.trace('cluster_service has instantiated');
-            context.services.execution = clusterService;
-            return require('./services/jobs')(context);
-        })
-        .then((jobsService) => {
-            logger.trace('jobs_service has instantiated');
-            context.services.jobs = jobsService;
-            return assetServiceOnline();
-        })
-        .then(() => require('./services/api')(context, app))
-        .then((apiService) => {
-            logger.trace('api_service has instantiated');
-            context.services.api = apiService;
-            return makeLogs(context);
-        })
-        .catch((err) => {
-            const errMsg = parseError(err);
-            logger.error('error during service initialization', errMsg);
-            messaging.send({ message: 'cluster:error:terminal' });
-        });
-
-    events.on('cluster_master:shutdown', () => {
+    messaging.register('worker:shutdown', () => {
         // Would there be sequencing required here?
         logger.debug('cluster_master is shutting down');
         const shuttingdown = _.map(context.services, service => service.shutdown());
@@ -98,4 +76,27 @@ module.exports = function (context) {
             }
         }));
     }
+    require('./services/execution.js')(context)
+        .then((clusterService) => {
+            messaging.initialize({ server });
+            logger.trace('cluster_service has instantiated');
+            context.services.execution = clusterService;
+            return require('./services/jobs')(context);
+        })
+        .then((jobsService) => {
+            logger.trace('jobs_service has instantiated');
+            context.services.jobs = jobsService;
+            return assetServiceOnline();
+        })
+        .then(() => require('./services/api')(context, app))
+        .then((apiService) => {
+            logger.trace('api_service has instantiated');
+            context.services.api = apiService;
+            return makeLogs(context);
+        })
+        .catch((err) => {
+            const errMsg = parseError(err);
+            logger.error('error during service initialization', errMsg);
+            messaging.send({ message: 'cluster:error:terminal' });
+        });
 };
