+++ lib/cluster/services/cluster/backends/native.js
@@ -5,47 +5,45 @@ const Promise = require('bluebird');
 const shortid = require('shortid');
 const Queue = require('queue');
 const parseError = require('../../../../utils/error_utils').parseError;
-const sendError = require('../../../../utils/api_utils').sendError;
 
-module.exports = function module(context) {
-    const messaging = context.messaging;
+/*
+ Execution Life Cycle for _status
+ pending -> scheduling -> running -> [ paused -> running ] -> [ stopped | completed ]
+ Exceptions
+ rejected - when a job is rejected prior to scheduling
+ failed - when there is an error while the job is running
+ aborted - when a job was running at the point when the cluster shutsdown
+ */
+
+module.exports = function module(context, messaging, executionService) {
     const events = context.foundation.getEventEmitter();
-    const logger = context.foundation.makeLogger({ module: 'cluster_service' });
-    // FIXME: this setting is for notifyNode
+    const logger = context.foundation.makeLogger({ module: 'native_service' });
     const configTimeout = context.sysconfig.teraslice.network_timeout;
     const pendingWorkerRequests = new Queue();
+    const nodeDisconnectTimeout = context.sysconfig.teraslice.node_disconnect_timeout;
+    const nodeStateInterval = context.sysconfig.teraslice.node_state_interval;
     const clusterState = {};
+    const clusterStats = {
+        slicer: {
+            processed: 0,
+            failed: 0,
+            queued: 0,
+            job_duration: 0,
+            workers_joined: 0,
+            workers_disconnected: 0,
+            workers_reconnected: 0
+        }
+    };
 
     // temporary holding spot used to attach nodes that are non responsive or
     // disconnect before final cleanup
     const droppedNodes = {};
 
-    // events can be fired from anything that instantiates a client, such as stores
-    events.on('getClient:config_error', _terminalShutdown);
-    events.on('jobs_service:verify_assets', _verifyAssets);
+    // TODO where does this go?
+    messaging.register('worker:shutdown', () => {
+        events.emit('cluster_master:shutdown');
+    });
 
-    // to catch signal propagation, but cleanup through msg sent from master
-    function _noOP() {
-    }
-
-    function _terminalShutdown(errEV) {
-        logger.error(`Terminal error: ${errEV.err}`);
-        messaging.send({ message: 'cluster:error:terminal' });
-    }
-
-    // TODO: see what this does
-    function _verifyAssets(assets) {
-        const msgObj = _.assign({ message: 'jobs_service:verify_assets' }, assets);
-        messaging.send(msgObj);
-    }
-
-    /*
-     * Should take note that since job_service is instantiated after cluster service,
-     * the mode to interact with job_service is through events emitters
-     */
-    // TODO: see if this is needed
-    messaging.register('process:SIGTERM', _noOP);
-    messaging.register('process:SIGINT', _noOP);
 
     messaging.register('node:online', 'node_id', (data, nodeId) => {
         logger.info(`node ${nodeId} has connected`);
@@ -67,6 +65,13 @@ module.exports = function module(context) {
         logger.debug(`node ${data.node_id} state is being updated`);
         logger.trace(`node ${data.node_id} has updated state:`, data);
         events.emit('cluster:available_workers');
+    });
+
+    // TODO this will become unneeded in message refactor
+    messaging.register('node:message:processed', (data) => {
+        // emitting the unique msg id to allow easier listener cleanup
+        logger.debug(`node message ${data._msgID} has been processed`);
+        events.emit(data._msgID, data);
     });
 
     messaging.register('node:workers:over_allocated', (requestData) => {
@@ -91,11 +96,10 @@ module.exports = function module(context) {
 
     messaging.register('job:error:terminal', (data) => {
         logger.debug(`terminal job error for execution: ${data.ex_id}, canceling job`);
+        const metaData = errorMetaData(data);
         pendingWorkerRequests.remove(data.ex_id);
         messaging.send('cluster:job:stop', { ex_id: data.ex_id });
-        if (!data.markedJob) {
-            events.emit('cluster:job_failure', data);
-        }
+        executionService.setExecutionStatus(data.ex_id, 'failed', metaData);
     });
 
     messaging.register('network:error', (err) => {
@@ -103,7 +107,7 @@ module.exports = function module(context) {
         logger.error('Error : cluster_master had an error with one of its connections', errMsg);
     });
 
-    messaging.register('network:disconnect', ['node_id', 'moderator'], (msg, id, identifier) => {
+    messaging.register('network:disconnect', ['node_id'], (msg, id, identifier) => {
         if (identifier) {
             // is a node_id
             if (identifier === 'node_id' && clusterState[id]) {
@@ -113,70 +117,63 @@ module.exports = function module(context) {
                     delete clusterState[nodeId];
                 } else {
                     clusterState[nodeId].state = 'disconnected';
-
                     const timer = setTimeout(() => {
                         _cleanUpNode(nodeId);
-                    }, 5000);
+                    }, nodeDisconnectTimeout);
 
                     droppedNodes[nodeId] = timer;
                 }
             }
-
-            // if (identifier === 'moderator') {
-            /* logger.warn(`Moderator ${id} has disconnected from cluster_master`);
-            delete moderator[id];
-            if (Object.keys(moderator).length === 0) {
-                moderator = null;
-            } */
-            // }
         } else {
             logger.error(`cluster_master got a disconnect event on a unidentified connection. might be a network/service communication issue, msg: ${msg}`);
         }
     });
 
+    messaging.register('cluster:analytics', (msg) => {
+        if (!clusterStats[msg.kind]) {
+            logger.warn(`unrecognized cluster stats: ${msg.kind}`);
+            return;
+        }
+        _.forOwn(msg.stats, (value, field) => {
+            if (clusterStats[msg.kind][field] !== undefined) {
+                clusterStats[msg.kind][field] += value;
+            }
+        });
+    });
 
     setInterval(() => {
         logger.trace('cluster_master requesting state update for all nodes');
         messaging.send('cluster:node:state');
-    }, 5000);
-
+    }, nodeStateInterval);
 
     function _cleanUpNode(nodeId) {
         // check workers and slicers
         const node = _checkNode(clusterState[nodeId]);
-
+        // if disconnected node had a slicer, we stop the execution of each slicer on it
+        // and mark it as failed
         if (node.hasSlicer) {
-            _.forIn(node.slicerJobID, (exId) => {
+            _.forIn(node.slicerExecutions, (exId) => {
                 logger.error(`node ${nodeId} has disconnected with active slicer for job: ${exId}, marking as failed and shutting down job`);
-
                 const data = { ex_id: exId, error: 'node_master where slicer resided has disconnected' };
+                const metaData = errorMetaData(data);
                 pendingWorkerRequests.remove(exId);
-                events.emit('cluster:job_failure', data);
+                executionService.setExecutionStatus(data.ex_id, 'failed', metaData);
                 messaging.send('cluster:job:stop', { ex_id: exId });
             });
         }
-
-        _.forIn(node.workerJobID, (__, exId) => {
+        // for any other worker not part of what is being shutdown, we attempt to reallocate
+        _.forIn(node.workerExecutions, (__, exId) => {
             // looking for unique ex_id's not in slicerJobID
-            if (!node.slicerJobID[exId]) {
-                // emitting to job service, since cluster is instantiated before
-                // job service is ready
-                const numOfWorkers = clusterState[nodeId]
-                    .active.filter(worker => worker.ex_id === exId).length;
-                // TODO: investigate the need for id
-                events.emit('cluster_service:cleanup_job', {
-                    ex_id: exId,
-                    id: exId,
-                    node_id: nodeId,
-                    workers: numOfWorkers,
-                    assignment: 'worker'
-                });
+            if (!node.slicerExecutions[exId]) {
+                const activeWorkers = clusterState[nodeId].active;
+                const numOfWorkers = activeWorkers.filter(worker => worker.ex_id === exId).length;
+                executionService.getActiveExecution(exId)
+                    .then(execution => addWorkers(execution, numOfWorkers));
             }
         });
 
         // cleanup key so we don't get ever growing obj
         delete droppedNodes[nodeId];
-
         delete clusterState[nodeId];
     }
 
@@ -184,12 +181,16 @@ module.exports = function module(context) {
         return _.cloneDeep(clusterState);
     }
 
-    function __checkNode(node) {
+    function getClusterStats() {
+        return _.cloneDeep(clusterStats);
+    }
+
+    function _checkNode(node) {
         const obj = {
             hasSlicer: false,
             numOfSlicers: 0,
-            slicerJobID: {},
-            workerJobID: {},
+            slicerExecutions: {},
+            workerExecutions: {},
             numOfWorkers: 0,
             id: node.id,
             available: node.available
@@ -199,16 +200,16 @@ module.exports = function module(context) {
             if (curr.assignment === 'slicer') {
                 prev.hasSlicer = true;
                 prev.numOfSlicers += 1;
-                prev.slicerJobID[curr.ex_id] = curr.ex_id;
+                prev.slicerExecutions[curr.ex_id] = curr.ex_id;
             }
 
             if (curr.assignment === 'worker') {
                 prev.numOfWorkers += 1;
                 // if not resgistered, set it to one, if so then increment it
-                if (!prev.workerJobID[curr.ex_id]) {
-                    prev.workerJobID[curr.ex_id] = 1;
+                if (!prev.workerExecutions[curr.ex_id]) {
+                    prev.workerExecutions[curr.ex_id] = 1;
                 } else {
-                    prev.workerJobID[curr.ex_id] += 1;
+                    prev.workerExecutions[curr.ex_id] += 1;
                 }
             }
 
@@ -220,7 +221,7 @@ module.exports = function module(context) {
         let slicerNode = null;
         for (let i = 0; i < stateArray.length; i += 1) {
             if (stateArray[i].state === 'connected' && stateArray[i].available > 0 && !errorNodes[stateArray[i].node_id]) {
-                const node = __checkNode(stateArray[i]);
+                const node = _checkNode(stateArray[i]);
 
                 if (!node.hasSlicer) {
                     slicerNode = stateArray[i].node_id;
@@ -238,7 +239,31 @@ module.exports = function module(context) {
         return slicerNode;
     }
 
-    // TODO: needReadyForAllocation => bool >= 2
+    function _findNodesForExecution(exId, slicerOnly) {
+        const nodes = [];
+        // TODO but what about disconnected nodes?
+        _.forOwn(clusterState, (node) => {
+            const hasJob = node.active.filter((worker) => {
+                if (slicerOnly) {
+                    return worker.ex_id === exId && worker.assignment === 'slicer';
+                }
+
+                return worker.ex_id === exId;
+            });
+
+            if (hasJob.length >= 1) {
+                nodes.push({
+                    node_id: node.node_id,
+                    ex_id: exId,
+                    hostname: node.hostname,
+                    workers: hasJob
+                });
+            }
+        });
+
+        return nodes;
+    }
+
     function _availableWorkers(all, forceCheck) {
         let num = 0;
         // determine which key to search for in cluster state
@@ -256,7 +281,7 @@ module.exports = function module(context) {
     }
 
     function _findPort(nodeId) {
-        return notifyNode(nodeId, 'cluster:node:get_port', {});
+        return _notifyNode(nodeId, 'cluster:node:get_port', {});
     }
 
     function _makeDispatch() {
@@ -275,11 +300,43 @@ module.exports = function module(context) {
         return methods;
     }
 
+    function _notifyNode(nodeId, msg, msgData) {
+        return new Promise(((resolve, reject) => {
+            // setting an unique id to know which message belongs to which
+            msgData._msgID = shortid.generate();
+            // set up the listener, since there is no clean way to have a
+            // dynamically named function or variable to identify the
+            // correct function to remove on cleanup, we are listening
+            // on and emitting the unique id as key
+            events.on(msgData._msgID, (nodeMasterData) => {
+                // remove listener to prevent memory leaks
+                events.removeAllListeners(msgData._msgID);
+                if (nodeMasterData.error) {
+                    reject(`Error: ${nodeMasterData.error} occured on node: ${nodeMasterData.node_id}`);
+                } else {
+                    logger.trace(`cluster_master completed message transaction with node: ${nodeId} for msg: ${msg} , msgData:`, msgData);
+                    resolve(nodeMasterData);
+                }
+            });
+            // send message
+            logger.debug(`cluster_master communicating with node: ${nodeId}`);
+            logger.trace(`cluster_master communicating with node: ${nodeId}, msg: ${msg}, msgData:`, msgData);
+            messaging.send(nodeId, msg, msgData);
+
+            // reject if timeout has been reached
+            setTimeout(() => {
+                // remove listener to prevent memory leaks
+                events.removeAllListeners(msgData._msgID);
+                reject(`timeout error while communicating with node: ${nodeId}, msg: ${msg}, data: ${JSON.stringify(msgData)}`);
+            }, configTimeout);
+        }));
+    }
+
     // designed to allocate additional workers, not any future slicers
-    function allocateWorkers(job, numOfWorkersRequested) {
-        const exId = job.ex_id;
-        const jobId = job.job_id;
-        const jobStr = JSON.stringify(job);
+    function allocateWorkers(execution, numOfWorkersRequested) {
+        const exId = execution.ex_id;
+        const jobId = execution.job_id;
+        const jobStr = JSON.stringify(execution);
         const sortedNodes = _.orderBy(clusterState, 'available', 'desc');
         let workersRequested = numOfWorkersRequested;
         let availWorkers = _availableWorkers(false, true);
@@ -330,7 +387,7 @@ module.exports = function module(context) {
             };
 
             results.push(
-                notifyNode(nodeId, 'cluster:workers:create', requestedWorkersData)
+                _notifyNode(nodeId, 'cluster:workers:create', requestedWorkersData)
                     .catch(() => {
                         logger.error(`An error has occurred in allocating : ${workerCount} workers to node : ${nodeId} , the worker request has been enqueued`);
                         pendingWorkerRequests.enqueue(requestedWorkersData);
@@ -368,7 +425,7 @@ module.exports = function module(context) {
                 recover_execution: recoverExecution
             };
 
-            return notifyNode(slicerNodeID, 'cluster:slicer:create', data);
+            return _notifyNode(slicerNodeID, 'cluster:slicer:create', data);
         });
     }
 
@@ -415,11 +472,11 @@ module.exports = function module(context) {
 
     events.on('cluster:available_workers', schedulePendingRequests);
 
-
     function shutdown() {
-        logger.info('shutting down.');
+        logger.info('shutting down');
         return Promise.resolve(true);
     }
+
 
     function _iterateState(cb) {
         return _.chain(clusterState)
@@ -436,14 +493,47 @@ module.exports = function module(context) {
             .flatten()
             .value();
     }
+    // TODO find out how to remove these
+    function findAllSlicers() {
+        return _iterateState(worker => worker.assignment === 'slicer');
+    }
 
-    // TODO: verify if this can stay
     function findWorkersByExecutionID(exId) {
         return _iterateState(worker => worker.assignment === 'worker' && worker.ex_id === exId);
     }
 
-    // FIXME: remove res from function, it should not be using http
-    function removeWorkers(res, exId, workerNum) {
+    function findSlicersByExecutionID(exId) {
+        return _iterateState(worker => worker.assignment === 'slicer' && worker.ex_id === exId);
+    }
+
+    function addWorkers(execution, workerNum) {
+        const workerData = {
+            job: JSON.stringify(execution),
+            id: execution.ex_id,
+            ex_id: execution.ex_id,
+            job_id: execution.job_id,
+            workers: workerNum,
+            assignment: 'worker'
+        };
+        pendingWorkerRequests.enqueue(workerData);
+        return { action: 'enqueued', ex_id: execution.ex_id, workerNum };
+    }
+
+    function setWorkers(execution, workerNum) {
+        const totalWorker = findWorkersByExecutionID(execution.ex_id).length;
+        if (totalWorker > workerNum) {
+            const removedWorkersCount = totalWorker - workerNum;
+            return removeWorkers(execution.ex_id, removedWorkersCount);
+        }
+        if (totalWorker < workerNum) {
+            const addWorkersCount = workerNum - totalWorker;
+            return addWorkers(execution, addWorkersCount);
+        }
+        // if they are equal then no work needs to be done
+        return { action: 'set', ex_id: execution.ex_id, workerNum };
+    }
+
+    function removeWorkers(exId, workerNum) {
         const dispatch = _makeDispatch();
         const workers = findWorkersByExecutionID(exId);
         let workerCount = workerNum;
@@ -457,42 +547,139 @@ module.exports = function module(context) {
 
             return prev;
         }, { _total: 0 });
-        // FIXME: msg does not exist, it should throw and api service should respond
-        if (workerNum >= workersData._total && msg === 'remove') {
-            sendError(res, 400, `workers to be removed: ${workerNum} cannot be >= to current workers: ${workersData._total}`);
-            return;
+        // TODO: test this error scenario
+        if (workerNum >= workersData._total) {
+            const errMsg = `workers to be removed: ${workerNum} cannot be >= to current workers: ${workersData._total}`;
+            logger.error(errMsg);
+            return Promise.reject(errMsg);
         }
-        // FIXME: pull out function from loop
-        while (workerCount) {
-            _.forOwn(workersData, (value, key) => {
-                if (key !== '_total') {
-                    if (workersData[key] >= 1 && workerCount > 0) {
-                        dispatch.set(key, 1);
-                        workersData[key] -= 1;
-                        workerCount -= 1;
-                    }
+        function stateForDispatch(__, key) {
+            if (key !== '_total') {
+                if (workersData[key] >= 1 && workerCount > 0) {
+                    dispatch.set(key, 1);
+                    workersData[key] -= 1;
+                    workerCount -= 1;
                 }
-            });
+            }
+        }
+
+        while (workerCount) {
+            _.forOwn(workersData, stateForDispatch);
         }
 
         const nodes = dispatch.getDispatch();
-        const results = _.map(nodes, (val, key) => notifyNode(key, 'cluster:workers:remove', {
-            workers: val,
-            ex_id: exId
-        }));
+        const results = _.map(nodes, (val, key) => _notifyNode(key, 'cluster:workers:remove', { workers: val, ex_id: exId }));
 
-        Promise.all(results)
-            .then(() => {
-                res.status(200).send(`${workerNum} workers have been released from job: ${exId}`);
-            })
+        return Promise.all(results)
+            .then(() => ({ action: 'remove', ex_id: exId, workerNum }))
             .catch((err) => {
                 const errMsg = parseError(err);
                 logger.error(`Error while releasing workers from job ${exId}, error: ${errMsg}`);
-                sendError(res, 500, errMsg);
+                return Promise.reject(errMsg);
             });
     }
 
-    const api = {
+    function _notifyNodesWithExecution(exId, message, slicerOnly, excludeNode) {
+        return new Promise(((resolve, reject) => {
+            const requests = [];
+            let nodes = _findNodesForExecution(exId, slicerOnly);
+            if (excludeNode) {
+                nodes = nodes.filter(node => node.hostname !== excludeNode);
+            }
+            nodes.forEach((node) => {
+                const messageNode = _notifyNode(node.node_id, message, {
+                    ex_id: exId
+                });
+
+                requests.push(messageNode);
+            });
+
+            return Promise.all(requests)
+                .then(() => {
+                    resolve(true);
+                })
+                .catch((err) => {
+                    const errMsg = parseError(err);
+                    logger.error('could not notify cluster', errMsg);
+                    reject(errMsg);
+                });
+        }));
+    }
+
+    function getSlicerStats(exId) {
+        return new Promise(((resolve, reject) => {
+            let list;
+            let timer;
+            const nodeQueries = [];
+
+            if (exId) {
+                list = findSlicersByExecutionID(exId);
+            } else {
+                list = findAllSlicers();
+            }
+            const numberOfCalls = list.length;
+
+            if (numberOfCalls === 0) {
+                if (exId) {
+                    reject({ message: `Could not find active slicer for ex_id: ${exId}`, code: 404 });
+                } else {
+                    // for the general slicer stats query, just return a empty array
+                    resolve([]);
+                }
+            }
+
+            _.each(list, (slicer) => {
+                const msg = { ex_id: slicer.ex_id };
+                nodeQueries.push(_notifyNode(slicer.node_id, 'cluster:slicer:analytics', msg));
+            });
+
+            Promise.all(nodeQueries)
+                .then((results) => {
+                    clearTimeout(timer);
+                    resolve(results.map(slicer => slicer.data));
+                })
+                .catch(err => reject({ message: parseError(err), code: 500 }));
+
+            timer = setTimeout(() => {
+                reject({ message: 'Timeout has occurred for query', code: 500 });
+            }, context.sysconfig.teraslice.network_timeout);
+        }));
+    }
+
+    function readyForAllocation() {
+        return _availableWorkers() >= 2;
+    }
+    // TODO: when and where should this be used?
+    function clusterAvailable() {}
+
+
+    function errorMetaData(data) {
+        const metaData = { _has_errors: 'true', _slicer_stats: data.slicer_stats };
+        logger.error(`job ${data.ex_id} has failed to complete`);
+
+        if (data.error) {
+            metaData._failureReason = data.error;
+        }
+        return metaData;
+    }
+
+    function stopExecution(exId) {
+        pendingWorkerRequests.remove(exId);
+        return _notifyNodesWithExecution(exId, 'cluster:job:stop', false, null);
+    }
+
+    function pauseExecution(exId) {
+        return _notifyNodesWithExecution(exId, 'cluster:job:pause', true, null);
+    }
+
+    function resumeExecution(exId) {
+        return _notifyNodesWithExecution(exId, 'cluster:job:resume', true, null);
+    }
+
+
+    // getClusterStats,
+
+    /* const api = {
         getClusterState,
         getSlicerStats,
         readyForAllocation,
@@ -505,16 +692,30 @@ module.exports = function module(context) {
         stopJob,
         pauseJob,
         resumeJob
+    }; */
+
+    const api = {
+        getClusterState,
+        getClusterStats,
+        getSlicerStats,
+        allocateWorkers,
+        allocateSlicer,
+        shutdown,
+        stopExecution,
+        pauseExecution,
+        resumeExecution,
+        removeWorkers,
+        addWorkers,
+        setWorkers,
+        readyForAllocation,
+        clusterAvailable
     };
 
     function _initialize() {
-        // TODO: handle the initialization of the cluster and only
-        // actually resolve the promise once everything is up and running.
-        // Should be delays here as we wait for nodes to join and share their
-        // state.
         logger.info('Initializing');
         return Promise.resolve(api);
     }
 
     return _initialize();
 };
+
