+++ lib/cluster/services/jobs.js
@@ -42,7 +42,8 @@ module.exports = function module(context) {
     }
 
     function updateJob(jobId, updatedJob) {
-        return _validateJob(updatedJob)
+        return _ensureAssets(updatedJob)
+            .then(parsedUpdatedJob => _validateJob(parsedUpdatedJob))
             .then(validJob => jobStore.update(jobId, validJob))
             .catch((err) => {
                 const errMsg = parseError(err);
